+++ src/StreamMemory.h
@@ -33,7 +33,7 @@ class CStreamMemory: public CStream
 {
 protected:
 	CByteArray m_Data;
-	int m_read_location;
+	size_t m_read_location;
 
 public:
 	CStreamMemory();
